login, manageOffences work, logOffence lowkey broken...will fix later
