+++ server/index.js
@@ -17,377 +17,21 @@ app.use(express.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 //------------------------------------------------------------login
-app.post("/Login", (req, res) => {
-  const setlgEmail = req.body.setlgEmail;
-  const setlgPassword = req.body.setlgPassword;
-  const sqlLgget =
-    "Select user_id, role from users where email = ? and password = ?";
-  db.query(sqlLgget, [setlgEmail, setlgPassword], (err, result) => {
-    if (!result?.[0]) {
-      //if there is no result
-      res.send("incorrect");
-    } else {
-      res.send(result); //send through data to store in session storage
-    }
-  });
-});
+const login=require("./login");// js file where code resides
+app.use('/', login);
 //-------------------------------------------------------------end login
 
-//log an offence when evidence hasn't been submitted (evidence optional)
-app.post("/LogOffenceNoFile", (req, res) => {
-  // post request to add a log
-  const offenderName = req.body.offenderName;
-  const offenceType = req.body.offenceType;
-  const offenceDetails = req.body.offenceDetails;
-  const offenceCode = req.body.offenceCode;
-  const offenceOther = req.body.offenceOther;
-  const submittedBy = req.body.submittedBy;
-  const offenceStatus = req.body.offenceStatus;
-  let offenceID;
-  let ticket_id;
-  console.log(req.body);
-  const sqlSelect =
-    "SELECT offence_id FROM offence_list WHERE offence_name = ?"; // get offence_id from table
-  db.query(sqlSelect, [offenceType], (err, result) => {
-    if (result[0] != null) {
-      offenceID = result[0].offence_id;
-      const sqlInsert =
-        "INSERT INTO logged_offences ( offender_name, offence_id, details, crs_code, offence_status, submitter_id ) VALUES (?,?,?,?,?,?);"; // insert into log table
-      db.query(
-        sqlInsert,
-        [
-          offenderName,
-          offenceID,
-          offenceDetails,
-          offenceCode,
-          offenceStatus,
-          submittedBy,
-        ],
-        (err, result) => {
-          const sqlGetId =
-            "SELECT * FROM logged_offences ORDER BY ticket_id DESC LIMIT 1"; // get ticket_id the ticket we just created
-          db.query(sqlGetId, (err, result) => {
-            ticket_id = result[0].ticket_id;
-            const sqlInsert =
-              "INSERT INTO other (ticket_id, offence_name) VALUES (?,?);"; // insert into log table
-            if (offenceType === "other") {
-              db.query(sqlInsert, [ticket_id, offenceOther], (err, result) => {
-                if (err !== null) {
-                  //if there is an error inserting into other, delete the offence from logged_offences
-                  const sqlDelete =
-                    "DELETE FROM logged_offences ORDER BY ticket_id DESC LIMIT 1";
-                  db.query(sqlDelete, (err, result));
-                  //res.send("Failed");
-                  //return;
-                }
-              });
-            }
-            //create directory to store all evidence related to this ticket
-            const dir = "./Uploads/Evidence/ticket" + ticket_id;
-            const saveLink = "/Uploads/Evidence/ticket" + ticket_id;
-            fs.mkdir(dir, (err) => {
-              if (err) {
-                throw err;
-              }
-            });
-            //now insert this directory into database
-            const sqlUpdateLink =
-              "Update logged_offences set ticket_link=? where ticket_id=?";
-            db.query(sqlUpdateLink, [saveLink, ticket_id], (err, result) => {
-              if (err != null) {
-                console.log(err);
-              }
-            });
 
-            const sqlGetStudent =
-              "select user_id from users where organization_nr=?"; //get student user id from database
-            const sqlInsertAction =
-              "Insert into actions (student_id, tables, table_id, seen, date, action_desc) values (?, ?, ?, ?, ?, ?)"; //this will allow the student to get an alert
-            db.query(sqlGetStudent, [offenderName], (err, result) => {
-              let student_id = result[0].user_id;
-              let table = "logged_offence"; //this will later allow us to redirect the notifiaction to appropriate page
-              let tableID = ticket_id;
-              let seen = "false";
-              let date = new Date().toISOString().slice(0, 10);
-              let actionDesc = "An offence has been logged against you.";
-              db.query(
-                sqlInsertAction,
-                [student_id, table, tableID, seen, date, actionDesc],
-                (err, result) => {
-                  if (err != null) {
-                    console.log(err);
-                  }
-                }
-              );
-            });
-          });
-        }
-      );
-    }
-  });
+//here are still errors!!!!!!!!!!
+//----------------------------Logging an offence (both with and without evidence)
+const logOffence=require("./logOffence");
+app.use('/', logOffence);
+//--------------------------end of logging an offence
 
-  //set up our email.
-  let transporter = nodemailer.createTransport({
-    service: "gmail",
-    auth: {
-      user: "sdteamoops@gmail.com",
-      pass: "SD1Team1OOPS!",
-    },
-    tls: {
-      rejectUnauthorised: false,
-    },
-  });
-
-  const sqlSelectEmail = "select email from users where organization_nr=?"; //get student email from database
-  db.query(sqlSelectEmail, [offenderName], (err, result) => {
-    offenderEmail = result[0].email;
-    let mailOptions = {
-      from: "sdteamoops@gmail.com",
-      to: offenderEmail,
-      subject: "Logged Offence",
-      text:
-        "This is an auto generated email.\nA student has reported an offence against you under the category of " +
-        offenceType +
-        ", an investigation into this case will follow.",
-    };
-    //send email to student to alert them that an offence has been logged against them
-    transporter.sendMail(mailOptions, function (err, success) {
-      if (err) {
-        console.log(err);
-        res.send("Unable to send email to offender");
-      } else {
-        console.log("Email sent to " + offenderEmail);
-        res.send("Successful");
-      }
-    });
-  });
-});
-
-//log an offence when a pdf file has been submitted as evidence
-app.post("/LogOffence", uploadEvidenceDoc.single("file"), (req, res) => {
-  // post request to add a log
-  const offenderName = req.body.offenderName;
-  const offenceType = req.body.offenceType;
-  const offenceDetails = req.body.offenceDetails;
-  const offenceCode = req.body.offenceCode;
-  const offenceOther = req.body.offenceOther;
-  const submittedBy = req.body.submittedBy;
-  const offenceStatus = req.body.offenceStatus;
-  let offenderEmail = req.body.offenderEmail;
-  let offenceID;
-  let ticket_id;
-  console.log(req.body);
-  const sqlSelect =
-    "SELECT offence_id FROM offence_list WHERE offence_name = ?"; // get offence_id from table
-  db.query(sqlSelect, [offenceType], (err, result) => {
-    if (result[0] != null) {
-      offenceID = result[0].offence_id;
-      const sqlInsert =
-        "INSERT INTO logged_offences ( offender_name, offence_id, details, crs_code, offence_status, submitter_id ) VALUES (?,?,?,?,?,?);"; // insert into log table
-      db.query(
-        sqlInsert,
-        [
-          offenderName,
-          offenceID,
-          offenceDetails,
-          offenceCode,
-          offenceStatus,
-          submittedBy,
-        ],
-        (err, result) => {
-          if (err != null) {
-            res.send("Failed");
-            return;
-          } else {
-            //create directory named for the ticket id
-            const sqlGetId =
-              "SELECT ticket_id FROM logged_offences ORDER BY ticket_id DESC LIMIT 1"; //get ticket
-            db.query(sqlGetId, (err, result) => {
-              ticket_id = result[0].ticket_id;
-              const dir = "./Uploads/Evidence/ticket" + ticket_id;
-              const saveLink = "/Uploads/Evidence/ticket" + ticket_id;
-              fs.mkdir(dir, (err) => {
-                if (err) {
-                  throw err;
-                }
-              });
-              //now insert this directory into database
-              const sqlUpdateLink =
-                "Update logged_offences set ticket_link=? where ticket_id=?";
-              db.query(sqlUpdateLink, [saveLink, ticket_id], (err, result) => {
-                if (err != null) {
-                  console.log(err);
-                }
-              });
-              //let fileType=req.file.mimetype.split("/")[1];
-
-              //get the file sent through and save it in the directory just created
-              let newFileName = Date.now() + req.file.originalname; //adding date ensures unique file name and lessens possibility of file loss and overwrites
-              let oldPath = "./Uploads/Evidence/" + req.file.filename; //file path where it has been uploaded
-              let newPath =
-                "./Uploads/Evidence/ticket" + ticket_id + "/" + newFileName; //created directory that we want to move the file to
-              fs.rename(oldPath, newPath, function (err) {
-                console.log(err);
-              });
-
-              const sqlInsert =
-                "INSERT INTO other (ticket_id, offence_name) VALUES (?,?);"; // insert into log table
-              if (offenceType === "other") {
-                db.query(
-                  sqlInsert,
-                  [ticket_id, offenceOther],
-                  (err, result) => {
-                    if (err !== null) {
-                      //if failed to insert into other, delete from logged offence
-                      const sqlDelete =
-                        "DELETE FROM logged_offences ORDER BY ticket_id DESC LIMIT 1";
-                      db.query(sqlDelete, (err, result));
-                      res.send("Failed");
-                      //return;
-                    }
-                  }
-                );
-              }
-              const sqlGetStudent =
-                "select user_id from users where organization_nr=?"; //get student user id from database
-              const sqlInsertAction =
-                "Insert into actions (student_id, tables, table_id, seen, date, action_desc) values (?, ?, ?, ?, ?, ?)"; //send action to student to get notification
-              db.query(sqlGetStudent, [offenderName], (err, result) => {
-                let student_id = result[0].user_id;
-                let table = "logged_offence";
-                let tableID = ticket_id;
-                let seen = "false";
-                let date = new Date().toISOString().slice(0, 10);
-                let actionDesc = "An offence has been logged against you.";
-                db.query(
-                  sqlInsertAction,
-                  [student_id, table, tableID, seen, date, actionDesc],
-                  (err, result) => {
-                    if (err != null) {
-                      console.log(err);
-                    }
-                  }
-                );
-              });
-            });
-          }
-        }
-      );
-    }
-  });
-
-  let transporter = nodemailer.createTransport({
-    service: "gmail",
-    auth: {
-      user: "sdteamoops@gmail.com",
-      pass: "SD1Team1OOPS!",
-    },
-    tls: {
-      rejectUnauthorised: false,
-    },
-  });
-
-  const sqlSelectEmail = "select email from users where organization_nr=?";
-  db.query(sqlSelectEmail, [offenderName], (err, result) => {
-    offenderEmail = result[0].email;
-    let mailOptions = {
-      from: "sdteamoops@gmail.com",
-      to: offenderEmail,
-      subject: "Logged Offence",
-      text:
-        "This is an auto generated email.\nA student has reported an offence against you under the category of " +
-        offenceType +
-        ", an investigation into this case will follow.",
-    };
-    transporter.sendMail(mailOptions, function (err, success) {
-      //send email to alert student that an offence has been logged against them
-      if (err) {
-        console.log(err);
-        res.send("Unable to send email to offender");
-      } else {
-        console.log("Email sent to " + offenderEmail);
-        res.send("Successful");
-      }
-    });
-  });
-});
-
-//get request to get list of possible offences, used to display clickable editable table
-app.get("/PossibleOffences", (req, res) => {
-  const sqlSelect =
-    "select offence_id,offence_name, severity ,offence_desc from offence_list where offence_name != 'other'";
-  db.query(sqlSelect, (error, result) => {
-    res.send(result);
-  });
-});
-
-//fetch the data from the database to send to frontend to show admin all the offences that have been logged
-app.get("/SubmittedOffences", (req, res) => {
-  const sqlSelect =
-    "select logged_offences.ticket_id, offender_name, (case when offence_list.offence_name='other' then other.offence_name else offence_list.offence_name end) as offence_name,crs_code , offence_status from logged_offences left join offence_list on logged_offences.offence_id= offence_list.offence_id left join other on logged_offences.ticket_id=other.ticket_id";
-  db.query(sqlSelect, (error, result) => {
-    res.send(result);
-  });
-});
-
-//do we still use this or is it the same as possible offences??
-app.get("/offences", (req, res) => {
-  const sqlSelect =
-    "select offence_id, offence_name from offence_list where offence_name != 'other'";
-  db.query(sqlSelect, (error, result) => {
-    res.send(result);
-  });
-});
-
-//post request to insert a new offence to the database
-app.post("/insert", (req, res) => {
-  const offenceName = req.body.offenceName;
-  const severity = req.body.severity;
-  const desc = req.body.desc;
-  const sqlSelect =
-    "Insert into offence_list(offence_name, severity, offence_desc) values(?,?,?)";
-  db.query(sqlSelect, [offenceName, severity, desc], (err, result) => {
-    console.log(err);
-    res.send("inserted");
-  });
-});
-
-//do we still use this??
-app.get("/selectOffence", (req, res) => {
-  const ticket_id = req.query["ticket_id"];
-  const sqlSelect = "select * from logged_offences where ticket_id = ?";
-  db.query(sqlSelect, [ticket_id], (err, result) => {
-    res.send(result[0]);
-  });
-});
-
-//query called when a possible offence get deleted
-app.post("/delete", (req, res) => {
-  const offenceId = req.body.offenceId;
-  const sqlSelect = "Delete from offence_list where offence_id = ?";
-  db.query(sqlSelect, [offenceId], (err, result) => {
-    if (err == null) {
-      res.send("deleted");
-    }
-  });
-});
-
-//query called a possible offence is edited
-app.post("/update", (req, res) => {
-  const offenceName = req.body.offenceName;
-  const severity = req.body.severity;
-  const desc = req.body.desc;
-  const offenceId = req.body.offenceId;
-  const sqlSelect =
-    "UPDATE offence_list SET severity = ?, offence_name =?, offence_desc=? WHERE offence_id = ?";
-  db.query(
-    sqlSelect,
-    [severity, offenceName, desc, offenceId],
-    (err, result) => {
-      res.send("updated");
-    }
-  );
-});
+//-------------------------managing offences (view possible offences,view submitted offences, add, edit, delete offences)
+const manageOffences=require("./manageOffences");
+app.use('/', manageOffences)
+//-----------------------end of managing offences
 
 //query to view the pledges.
 app.get("/viewPledges", (req, res) => {
